requirements.txt
pandas
openpyxl
pyyaml
oracledb